+++ logdissect/parsers/__init__.py
@@ -21,11 +21,11 @@
 # SOFTWARE.
 
 __all__ = ['syslogbsd', 'syslogiso', 'nohost', 'ldjson', 'tcpdump',
-        'cisco_ios']
+        'ciscoios']
 
 import logdissect.parsers.syslogbsd
 import logdissect.parsers.syslogiso
 import logdissect.parsers.nohost
 import logdissect.parsers.ldjson
 import logdissect.parsers.tcpdump
-import logdissect.parsers.cisco_ios
+import logdissect.parsers.ciscoios